# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Margaz Tank Telemetry API
  version: 1.0.0
  description: "API documentation for the Margaz Tank Telemetry backend. Frontend: https://margaz.netlify.app"
  license:
    name: Proprietary
    url: https://margaz.netlify.app
servers:
  - url: http://63.181.47.189
    description: Lightsail direct
...
security:
  - bearerAuth: []
paths:
...
  /api/dealers:
    get:
      tags: [Dealers]
      summary: List dealers
      operationId: listDealers
      security: []
      responses:
        '200':
          description: Dealer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dealer'
              example:
                - id: "84f99a90-7c47-47ad-9740-57235997d536"
                  title: "Aktup Test"
                  deviceId: "2-aktup"
                  tankLevel: 60
                  lastData: "2026-01-30T10:35:51.986Z"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Dealers]
      summary: Create dealer (Admin only)
      operationId: createDealer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerCreateRequest'
            example:
              title: "Yeni Bayi"
              city: "KAYSERI"
              district: "DEVELI"
              address: "Ornek Mah. No:1"
              deviceId: "2-aktup"
              tankLevel: 55
      responses:
        '201':
          description: Dealer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/dealers/{id}:
    get:
      tags: [Dealers]
      summary: Get dealer by id
      operationId: getDealer
      security: []
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
              example:
                id: "84f99a90-7c47-47ad-9740-57235997d536"
                title: "Aktup Test"
                deviceId: "2-aktup"
                tankLevel: 60
                lastData: "2026-01-30T10:35:51.986Z"
        '404':
          description: Dealer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Dealers]
      summary: Update dealer (Admin only)
      operationId: updateDealer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DealerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerUpdateRequest'
            example:
              title: "Aktup Test"
              deviceId: "2-aktup"
              tankLevel: 65
      responses:
        '200':
          description: Dealer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Dealer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Dealers]
      summary: Delete dealer (Admin only)
      operationId: deleteDealer
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DealerId'
      responses:
        '200':
          description: Dealer deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Dealer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/dealers/{id}/history:
    get:
      tags: [Dealers]
      summary: Get dealer telemetry history
      operationId: getDealerHistory
      security: []
      parameters:
        - $ref: '#/components/parameters/DealerId'
        - name: hours
          in: query
          required: false
          schema:
            type: integer
            default: 24
            minimum: 1
          description: History window in hours
      responses:
        '200':
          description: Telemetry history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryHistory'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/dealers/geocode:
    post:
      tags: [Dealers]
      summary: Trigger geocoding process (Admin only)
      operationId: geocodeDealers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Geocoding started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DealerId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    TelemetryHistory:
      type: object
      properties:
        tankLevel:
          type: number
        timestamp:
          type: string
          format: date-time
      required: [tankLevel, timestamp]
    Dealer:
      type: object
      properties:
        id:
          type: string
        licenseNo:
          type: string
        deviceId:
          type: string
          nullable: true
        title:
          type: string
        city:
          type: string
        district:
          type: string
        address:
          type: string
        status:
          type: string
        distributor:
          type: string
        taxNo:
          type: string
          nullable: true
        decisionNo:
          type: string
          nullable: true
        documentNo:
          type: string
          nullable: true
        contractStartDate:
          type: string
          format: date-time
          nullable: true
        contractEndDate:
          type: string
          format: date-time
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        tankLevel:
          type: number
        lastData:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title]
      example:
        id: "84f99a90-7c47-47ad-9740-57235997d536"
        title: "Aktup Test"
        deviceId: "2-aktup"
        tankLevel: 60
        lastData: "2026-01-30T10:35:51.986Z"
    DealerCreateRequest:
      type: object
      properties:
        licenseNo:
          type: string
        title:
          type: string
        city:
          type: string
        district:
          type: string
        address:
          type: string
        status:
          type: string
        distributor:
          type: string
        deviceId:
          type: string
          nullable: true
        tankLevel:
          type: number
          default: 0
      required: [title]
    DealerUpdateRequest:
      type: object
      properties:
        title:
          type: string
        city:
          type: string
        district:
          type: string
        address:
          type: string
        status:
          type: string
        distributor:
          type: string
        deviceId:
          type: string
          nullable: true
        tankLevel:
          type: number
...